src/app_transport_framework_library/pyproject.toml
[tool.poetry]
name = "app-transport-framework-library"
version = "0.1.12"
description = ""
authors = ["Sven Sommer <rob.hoffmann@posteo.de>"]
readme = "README.md"
packages = [{include = "app_transport_framework_library"}]

[tool.poetry.dependencies]
python = "^3.8"
fhir-resources = "^7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
